+++ test/test_core/test_headers_assertion.py
@@ -14,11 +14,11 @@ def test_check_not_raise_when_value_is_none(fake_assertion_data: AssertionData):
     checker = _HeadersChecker(value=None)
     with does_not_raise():
         # negative = False
-        checker.check(assertion_data=fake_assertion_data, negative=False)
+        assert evolve(fake_assertion_data, negative_assertion=False) in checker
 
     with does_not_raise():
         # negative = True
-        checker.check(assertion_data=fake_assertion_data, negative=True)
+        assert evolve(fake_assertion_data, negative_assertion=True) in checker
 
 
 @pytest.mark.parametrize(
@@ -37,10 +37,12 @@ def test_check_not_raise_when_value_is_none(fake_assertion_data: AssertionData):
     ],
 )
 def test_check_not_raise(should_not_raise: Spec, fake_assertion_data: AssertionData, response_headers, check_headers):
-    assertion_data = evolve(fake_assertion_data, response_headers=response_headers)
     checker = _HeadersChecker(value=check_headers)
     with should_not_raise.expected:
-        checker.check(assertion_data=assertion_data, negative=should_not_raise.negative)
+        assert (
+            evolve(fake_assertion_data, response_headers=response_headers, negative_assertion=should_not_raise.negative)
+            in checker
+        )
 
 
 @pytest.mark.parametrize(
@@ -61,7 +63,9 @@ def test_check_not_raise(should_not_raise: Spec, fake_assertion_data: AssertionD
 def test_check_when_header_not_match_response(
     should_raise: Spec, fake_assertion_data: AssertionData, response_headers, check_headers
 ):
-    assertion_data = evolve(fake_assertion_data, response_headers=response_headers)
     checker = _HeadersChecker(value=check_headers)
     with should_raise.expected:
-        checker.check(assertion_data=assertion_data, negative=should_raise.negative)
+        assert (
+            evolve(fake_assertion_data, response_headers=response_headers, negative_assertion=should_raise.negative)
+            in checker
+        )