+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 ## 0.8.0 (Unreleased)
 - chore: update dependencies
+- feat: improve assert report
 
 ## 0.7.0
 - feat: possible to pass string when check content